实验二/密码加密/requirements.txt
PyQt5>=5.15.0
pycryptodome>=3.15.0
gmssl>=3.2.1
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.0
Pillow>=8.0.0